+++ tests/test_bids.py
@@ -89,8 +89,8 @@ def test_create_participant_tsv(tmp_dir):
     assert df["participant_id"].values[0] == "sub-01"
 
 
-def test_create_sessions_tsv(tmp_dir):
-    """Test for ``create_sessions_tsv``."""
+"""def test_create_sessions_tsv(tmp_dir):
+    ""Test for ``create_sessions_tsv``.""
     from nifti2bids.simulate import simulate_bids_dataset
 
     path = simulate_bids_dataset(n_sessions=3, output_dir=Path(tmp_dir.name) / "BIDS")
@@ -102,7 +102,7 @@ def test_create_sessions_tsv(tmp_dir):
     assert filename.is_file()
 
     df = pd.read_csv(filename, sep="\t")
-    assert df["session_id"].values.tolist() == ["ses-1", "ses-2", "ses-3"]
+    assert df["session_id"].values.tolist() == ["ses-1", "ses-2", "ses-3"]"""
 
 
 def test_get_entity_value():
@@ -169,13 +169,13 @@ def test_PresentationBlockExtractor(tmp_dir, scanner_start_time, sum_duration_co
 
     extractor = PresentationBlockExtractor(
         log_or_df=filename,
-        block_cue_names=["indoor"],
+        block_cue_names=["indo.*"],
         convert_to_seconds=["Time", "Duration"],
         scanner_event_type="Pulse",
         scanner_trigger_code="99",
         n_discarded_volumes=1 if scanner_start_time is None else 0,
         tr=1 if scanner_start_time is None else None,
-        rest_block_code="rest",
+        rest_block_codes="res.*",
         quit_code="quit",
     )
 
@@ -201,7 +201,7 @@ def test_PresentationBlockExtractor_mean_rt_and_accuracy(tmp_dir):
         convert_to_seconds=["Time"],
         scanner_event_type="Pulse",
         scanner_trigger_code="99",
-        rest_block_code="rest",
+        rest_block_codes="rest",
         quit_code="quit",
         split_cue_as_instruction=True,
         drop_instruction_cues=True,
@@ -244,7 +244,7 @@ def test_PresentationBlockExtractor_instruction_cue_separation(
         convert_to_seconds=["Time", "Duration"],
         scanner_event_type="Pulse",
         scanner_trigger_code="99",
-        rest_block_code="rest",
+        rest_block_codes="rest",
         rest_code_frequency="variable",
         split_cue_as_instruction=True,
         quit_code="quit",
@@ -309,7 +309,7 @@ def test_PresentationEventExtractor(tmp_dir, scanner_start_time):
 
     extractor = PresentationEventExtractor(
         log_or_df=filename,
-        trial_types=["incongruentright", "congruentleft"],
+        trial_types=["incongruent.*", "congruentleft"],
         convert_to_seconds=["Time", "Duration"],
         scanner_event_type="Pulse",
         scanner_trigger_code="99",
@@ -371,7 +371,7 @@ def test_EPrimeBlockExtractor(tmp_dir, offset_column_name):
             convert_to_seconds=["Data.OnsetTime", "TriggerStart", "Data.OffsetTime"],
             n_discarded_volumes=1 if column else 0,
             tr=1 if column else None,
-            rest_block_code="Rest",
+            rest_block_codes="Rest",
             rest_code_frequency="variable",
         )
 
@@ -393,12 +393,12 @@ def test_EPrimeBlockExtractor_mean_rt_and_accuracy(tmp_dir):
 
     extractor = EPrimeBlockExtractor(
         log_or_df=filename,
-        block_cue_names=["A", "B"],
+        block_cue_names=["A.*", "B"],
         onset_column_name="Data.OnsetTime",
         procedure_column_name="Procedure",
         trigger_column_name="TriggerStart",
         convert_to_seconds=["Data.OnsetTime", "Data.RT", "TriggerStart"],
-        rest_block_code="Rest",
+        rest_block_codes="Res.*",
         rest_code_frequency="variable",
     )
 
@@ -452,7 +452,7 @@ def test_EPrimeBlockExtractor_instruction_cue_separation(
             "TriggerStart",
             "Data.OffsetTime",
         ],
-        rest_block_code="Rest",
+        rest_block_codes="Rest",
         rest_code_frequency="variable",
         split_cue_as_instruction=True,
         block_cues_without_instruction=("B"),